+++ src/pages/Login/LoginForm.tsx
@@ -1,0 +1,87 @@
+import { Field, Form, FormikErrors, FormikProps, withFormik } from 'formik'
+
+import styles from '@components/Authentication/Form.module.css'
+import { signIn } from '@utils/signIn'
+import { Link } from 'react-router-dom'
+import { Authentication } from '@components/Authentication/Authentication'
+
+interface LoginValues {
+  email: string
+  password: string
+}
+
+interface OtherProps {
+  message: string
+}
+
+interface IFormProps {
+  initialEmail?: string
+  message: string
+}
+
+const InnerForm = (props: OtherProps & FormikProps<LoginValues>) => {
+  const { touched, errors, isSubmitting, message } = props
+
+  return (
+    <Authentication>
+      <Form className={styles.loginForm}>
+        <h1 className={styles.headerTitle}>{message}</h1>
+        <div className={styles.formContainer}>
+          <div>
+            {touched.email && errors.email && <div className={styles.error}>{errors.email}</div>}
+            <Field
+              type='email'
+              name='email'
+              className={styles.form}
+              style={{ borderColor: touched.email && errors.email && 'red' }}
+              placeholder='Почта'
+            />
+          </div>
+
+          <div>
+            <Field type='password' name='password' className={styles.form} placeholder='Пароль' />
+            {touched.password && errors.password && <div>{errors.password}</div>}
+          </div>
+        </div>
+
+        <div>
+          <p>Нет аккаунта?</p>
+          <Link to={'/login'}>
+            <p>Создать аккаунт</p>
+          </Link>
+        </div>
+
+        <button type='submit' disabled={isSubmitting} className={styles.button}>
+          Войти
+        </button>
+      </Form>
+    </Authentication>
+  )
+}
+
+export const LoginForm = withFormik<IFormProps, LoginValues>({
+  mapPropsToValues: (props) => {
+    return {
+      email: props.initialEmail || '',
+      password: '',
+    }
+  },
+  validate: (values: LoginValues) => {
+    let errors: FormikErrors<LoginValues> = {}
+    if (!values.email) {
+      errors.email = 'Поле не может быть пустым'
+    } else if (!values.email) {
+      errors.email = 'Invalid email address'
+    }
+    return errors
+  },
+
+  handleSubmit: (values) => {
+    console.log(1)
+    try {
+      console.log(signIn(values.email, values.password))
+    } catch (error) {
+      console.log(error)
+    }
+  },
+})(InnerForm)
